feat: change redirect after login

+++ board/views.py
@@ -40,16 +40,14 @@ def topic(request, slug):
     return render(request, 'topic/index.html', context)
     
 def delete_topic(request, id):
-    topic = Topic.objects.get(id=id)
-    topic.delete()
+    if request.user.is_authenticated:
+        topic = Topic.objects.get(id=id)
+        if request.user == topic.user:
+            topic.delete()
+            return redirect('board:board')
     return redirect('board:board')
 
 def delete_post(request, topic_id, message_id):
     post = Post.objects.get(id=message_id)
     post.delete()
     return redirect('board:topic', id=topic_id)
-
-def delete_topic(request, id):
-    topic = Topic.objects.get(id=id)
-    topic.delete()
-    return redirect('board:board')
